Add files via upload
World Happiness Report is a dataset where the overall happiness of a country is calculated based on six factors. Various governments, organisations and other institutional bodies use this report to make sound decisions on their policies. The overall happiness depends upon the six factors which were learned during this project. to compare the happiness, there is a hypothetical country 'Dystopia' with the least happiness possible so that all the countries can be compared with it.