Add files via upload
Mushroom Class Prediction Project - In this project, we have to find out the class of a mushroom from Poisonous or Edible. The mushroom can be poisonous and it would be fatal to eat it, so the model needs to be the best. Here, I tried different classification models and compared all their accuracies. Using the cross-validation technique, I checked for the underfitting or overfilling of the model. I found out that the Decision Tree Classifier is the best possible model with the highest accuracy. Then, with the help of hyper-parameter tuning, I found out the best parameters for the Decision Tre Classifier. Finally, the fitted model was able to give an accuracy and f1-score of 100%.