Add files via upload
In this problem, the dataset utilizes data from 2014 Major League Baseball seasons in order to develop an algorithm that predicts the number of wins for a given team in the 2015 season based on several different indicators of success.
I have performed the complete EDA process along with the data visualisation and data cleaning wherever necessary. Then, select the best prediction model out of 7 different models.
Complete process and insights from the model are given in the notebook.