Add files via upload
In this case study, we needed to build the model for the prediction of Attrition in HR. We are given a dataset and we had to predict if a person will leave the company or not.

Many attributes of the data were categorical and they were encoded using label encoding, frequency encoding and one hot encoding. No outliers were found in the data during the EDA process. The scaling of the data was done using Min-Max scaler before using the data for model building.

7 models, we selected for the model building namely Logistic Regressor, Decision Tree Classifier, KNN Classifier, SVC Classifier, Random Forest Classifier, Adaboost Classifier, and MultinomialNB. These models are cross-validated to look for the best possible model with the highest mean accuracy. Two models Decision Tree and Random Forest were selected for the tuning after cross-validation.

After the tuning, Model Evaluation was done using classification report, confusion matrix and ROC curve. Random Forest gave the best performance with an AUC score of 0.99 and an f1 score of 0.94. It is then saved as an object to use for prediction.