Add files via upload
This is the notebook for the case study of Avacado Average Price prediction in the different cities of the USA. This case study uses a data set of dimension 1517 different indexes and 13 features including the target class. 
Complete steps including EDA, data visualization are shown in the notebook. The final model selected is Random Forest Classifier with an accuracy of 83%, mean absolute error of the model is 0.057 and the root mean squared error is 0.074.