Add files via upload
In this case study, we needed to build the model for the prediction of the income range of the Census. We are given a dataset and we had to predict if a person has income greater than $50K/year or less than $50K/year.

Many attributes of the data were categorical and they were encoded using label encoding, frequency encoding and one hot encoding. Some outliers were present in the data and they were removed during the EDA process. The scaling of the data was done using Min-Max scaler before using the data for model building.

 Out of 7 models, namely Logistic Regressor, Decision Tree Classifier, KNN Classifier, SVC Classifier, Random Forest Classifier, Adaboost Classifier, and MultinomialNB, Random Forest classifier is selected as the best model after comparing the mean accuracies from cross-validation and accuracy.

After the hyperparameter tuning, Model Evaluation was done using classification report, confusion matrix and ROC curve. Random Forest gave the best performance with an AUC score of 0.97 and an f1 score of 0.91. It is then saved as an object to use for prediction.