Add files via upload
This is a problem case study for the prediction of Loan Staus of the Loan Applicants. We have to predict whether a person's loan application will be approved or not by building a classification model. I have analysed the data set will the necessary features and target variable.

The best model selected was the Random Forest Classifier Model with an AUC score of 0.95 and an accuracy of 89%. Find the jupyter notebook named 'Loan Status Prediction Model.ipynb'.