Add files via upload
Flight ticket prices can be something hard to guess, today we might see a price, check out the price of the same flight tomorrow, it will be a different story. We might have often heard travellers saying that flight ticket prices are so unpredictable. In this case study, I predict the prices of flight tickets for various airlines between March and June of 2019 and between various cities.

I have performed a thorough analysis on what are the factors that affect the fare of flights throughout the given time period using EDA and data visualisation. Then, I built the regression model, which is selected after the cross-validation, tuning, and evaluation of 8 other models.

Please look at the file named 'Flight Price Prediction Project.ipynb'