Add files via upload
Insurance fraud is a huge problem in the industry. It's difficult to identify fraud claims. Machine Learning is in a unique position to help the Auto Insurance industry with this problem.

In this project, I used a dataset that has the details of the insurance policy along with the customer details. It also has the details of the accident on the basis of which the claims have been made.

I worked with some auto insurance data to demonstrate how you can create a predictive model that predicts if an insurance claim is fraudulent or not.

Please look at the notebook named 'Insurance Claim Fraud Prediction.ipynb'