Add files via upload
Problem Statement : 

One of the largest retail chains in the world wants to use its vast data source to build an efficient forecasting model to predict the sales for each SKU(stock-keeping unit) in its portfolio at its 76 different stores using historical sales data for the past 3 years on a week-on-week basis. Sales and promotional information is also available for each week - product and store wise. 

However, no other information regarding stores and products are available. So we need to forecast accurately the sales values for every such product/SKU-store combination for the next 12 weeks accurately.