Add files via upload
This pdf contains my handwritten notes on the basic theory of Linear Regression. I have tried to cover all the basics and concepts behind the Linear Regression Model. Gradient Descent and its concepts are also mentioned in this pdf.
I have also discussed the different assumptions that we take during working with a Linear Regression Model.
Please do let me know if I can improve on the concepts and theory shared in this pdf. Feel free to use it for your learning purpose only.